out of credits dialog
